requirements.txt
bcrypt==4.3.0
blinker==1.9.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
reportlab==4.3.1
Werkzeug==3.1.3
WTForms==3.2.1
gunicorn==20.1.0
